feat: Migration to Latest Dev Tools
- Upgrades gradle & gradle wrapper
- Updates the compile & target SDK to fix warnings
- Moves from 'appcompat' to 'androidx' library usage
- Moves hard coded strings to 'Strings.xml'
- Removes unused resources
- Fixes Lint warnings

This commit migrates the code base to the latest android development tools before adding new features.

The migration is done as follows,

- Analyze the project using lint and apply auto-suggesions wherever possible
- When auto-suggestion isn't available, apply sensible changes manually
