+++ app/src/main/java/com/android/nikhil/memorygame/CardAdapter.kt
@@ -1,7 +1,7 @@
 package com.android.nikhil.memorygame
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
