+++ app/src/main/java/com/android/nikhil/memorygame/GameCallback.kt
@@ -1,0 +1,6 @@
+package com.android.nikhil.memorygame
+
+interface GameCallback {
+    fun onWin()
+    fun onLose()
+}
