+++ app/src/main/res/layout/card_back.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -12,15 +12,16 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:padding="30dp"
         android:background="@drawable/gradient"
-        android:foreground="?selectableItemBackground">
+        android:foreground="?selectableItemBackground"
+        android:padding="30dp">
 
         <ImageView
             android:id="@+id/cardImageView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:src="@drawable/ic_fingerprint"/>
+            android:contentDescription="@string/card_back"
+            android:src="@drawable/ic_fingerprint" />
 
     </RelativeLayout>
-</android.support.v7.widget.CardView>
+</androidx.cardview.widget.CardView>
