+++ app/src/test/java/com/android/nikhil/memorygame/ExampleUnitTest.kt
@@ -1,0 +1,18 @@
+package com.android.nikhil.memorygame
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see [Testing documentation](http://d.android.com/tools/testing)
+ */
+class ExampleUnitTest {
+    @Test
+    @Throws(Exception::class)
+    fun addition_isCorrect() {
+        assertEquals(4, (2 + 2).toLong())
+    }
+}
